Move hinting of Templates to separate hook
